Moved relevant methods from views to either Formats/Group.py or Formats/Query.py.
I'd wanted it arranged differently (all qry-related methods in Query), but due to some circular imports, I had to put some in Group.py.  Teasing those apart better will be a more involved refactor.  For example, "SearchGroup.py" should not be imported in Query.py.  It's too specific.  So I need to make Query into a proper class.  There are also some methods that contain both Group.py method calls and Query.py method calls.  I think either those need to be more granularized and teased apart or I should re-think the separation.

Also, I noted that things like performQuery should probably be in Format.py, not Group.py, but the qry object definitely is structured for the group.  I need to think about this more...

Files checked in: DataRepo/Formats/Group.py DataRepo/Formats/Query.py DataRepo/templatetags/customtags.py DataRepo/tests/test_views.py DataRepo/views.py
