+++ DataRepo/Formats/Query.py
@@ -1,5 +1,7 @@
 from copy import deepcopy
 
+from django.http import Http404
+
 
 def extractFldPaths(qry):
     """
@@ -127,3 +129,305 @@ def appendFilterToGroup(parent, filter):
     output_filter = deepcopy(parent)
     output_filter["queryGroup"].append(filter)
     return output_filter
+
+
+def createNewBasicQuery(basv_metadata, mdl, fld, cmp, val, fmt):
+    """
+    Constructs a new qry object for an advanced search from basic search input.
+    """
+
+    qry = basv_metadata.getRootGroup(fmt)
+
+    try:
+        mdl = basv_metadata.getModelInstance(fmt, mdl)
+    except KeyError as ke:
+        raise Http404(ke)
+
+    sfields = basv_metadata.getSearchFields(fmt, mdl)
+
+    if fld not in sfields:
+        raise Http404(
+            f"Field [{fld}] is not searchable.  Must be one of [{','.join(sfields.keys())}]."
+        )
+
+    num_empties = basv_metadata.getNumEmptyQueries(qry["searches"][fmt]["tree"])
+    if num_empties != 1:
+        raise Http404(
+            f"The static filter for format {fmt} is improperly configured. It must contain exactly 1 empty query."
+        )
+
+    empty_qry = getFirstEmptyQuery(qry["searches"][fmt]["tree"])
+
+    empty_qry["type"] = "query"
+    empty_qry["pos"] = ""
+    empty_qry["static"] = False
+    empty_qry["fld"] = sfields[fld]
+    empty_qry["ncmp"] = cmp
+    empty_qry["val"] = val
+
+    dfld, dval = basv_metadata.searchFieldToDisplayField(mdl, fld, val, qry)
+
+    if dfld != fld:
+        # Set the field path for the display field
+        empty_qry["fld"] = sfields[dfld]
+        empty_qry["val"] = dval
+
+    return qry
+
+
+def getFirstEmptyQuery(qry_ref):
+    """
+    This method takes the "tree" from a qry object (i.e. what you get from basv_metadata.getRootGroup(fmt)) and returns
+    a reference to the single empty item of type query that should be present in a new rootGroup.
+    """
+    if qry_ref["type"] and qry_ref["type"] == "query":
+        if qry_ref["val"] == "":
+            return qry_ref
+        return None
+    elif qry_ref["type"] and qry_ref["type"] == "group":
+        immutable = qry_ref["static"]
+        if len(qry_ref["queryGroup"]) > 0:
+            for qry in qry_ref["queryGroup"]:
+                emptyqry = getFirstEmptyQuery(qry)
+                if emptyqry:
+                    if immutable:
+                        raise Http404(
+                            "Group containing empty query must not be static."
+                        )
+                    return emptyqry
+        return None
+    raise Http404("Type not found.")
+
+
+def isQryObjValid(qry, form_class_list):
+    """
+    Determines if an advanced search qry object was properly constructed/populated (only at the root).
+    """
+
+    if (
+        type(qry) is dict
+        and "selectedtemplate" in qry
+        and "searches" in qry
+        and len(form_class_list) == len(qry["searches"].keys())
+    ):
+        for key in form_class_list:
+            if (
+                key not in qry["searches"]
+                or type(qry["searches"][key]) is not dict
+                or "tree" not in qry["searches"][key]
+                or "name" not in qry["searches"][key]
+            ):
+                return False
+        return True
+    else:
+        return False
+
+
+def isValidQryObjPopulated(qry):
+    """
+    Checks whether a query object is fully populated with at least 1 search term.
+    """
+    selfmt = qry["selectedtemplate"]
+    if len(qry["searches"][selfmt]["tree"]["queryGroup"]) == 0:
+        return False
+    else:
+        return isValidQryObjPopulatedHelper(
+            qry["searches"][selfmt]["tree"]["queryGroup"]
+        )
+
+
+def isValidQryObjPopulatedHelper(group):
+    for query in group:
+        if query["type"] == "query":
+            if not query["val"] or query["val"] == "":
+                return False
+        elif query["type"] == "group":
+            if len(query["queryGroup"]) == 0:
+                return False
+            else:
+                tmp_populated = isValidQryObjPopulatedHelper(query["queryGroup"])
+                if not tmp_populated:
+                    return False
+    return True
+
+
+def formsetsToDict(rawformset, form_classes):
+    """
+    Takes a series of forms and a list of form fields and uses the pos field to construct a hierarchical qry tree.
+    """
+
+    # All forms of each type are all submitted together in a single submission and are duplicated in the rawformset
+    # dict.  We only need 1 copy to get all the data, so we will arbitrarily use the first one
+
+    # Figure out which form class processed the forms (inferred by the presence of 'saved_data' - this is also the
+    # selected format)
+    processed_formkey = None
+    for key in rawformset.keys():
+        # We need to identify the form class that processed the form to infer the selected output format.  We do that
+        # by checking the dictionary of each form class's first form for evidence that it processed the forms, i.e. the
+        # presence of the "saved_data" class data member which is created upon processing.
+        if "saved_data" in rawformset[key][0].__dict__:
+            processed_formkey = key
+            break
+
+    # If we were unable to locate the selected output format (i.e. the copy of the formsets that were processed)
+    if processed_formkey is None:
+        raise Http404(
+            f"Unable to find the saved form-processed data among formats: {','.join(rawformset.keys())}."
+        )
+
+    return formsetToDict(rawformset[processed_formkey], form_classes)
+
+
+def formsetToDict(rawformset, form_classes):
+    """
+    Helper for formsetsToDict that handles only the forms belonging to the selected output format.
+    """
+
+    search = {"selectedtemplate": "", "searches": {}}
+
+    # We take a raw form instead of cleaned_data so that form_invalid will repopulate the bad form as-is
+    isRaw = False
+    try:
+        formset = rawformset.cleaned_data
+    except AttributeError:
+        isRaw = True
+        formset = rawformset
+
+    for rawform in formset:
+
+        if isRaw:
+            form = rawform.saved_data
+        else:
+            form = rawform
+
+        path = form["pos"].split(".")
+
+        [format, formatName, selected] = rootToFormatInfo(path.pop(0))
+        rootinfo = path.pop(0)
+
+        # If this format has not yet been initialized
+        if format not in search["searches"]:
+            search["searches"][format] = {}
+            search["searches"][format]["tree"] = {}
+            search["searches"][format]["name"] = formatName
+
+            # Initialize the root of the tree
+            [pos, gtype, static] = pathStepToPosGroupType(rootinfo)
+            aroot = search["searches"][format]["tree"]
+            aroot["pos"] = ""
+            aroot["type"] = "group"
+            aroot["val"] = gtype
+            aroot["static"] = static
+            aroot["queryGroup"] = []
+            curqry = aroot["queryGroup"]
+        else:
+            # The root already exists, so go directly to its child list
+            curqry = search["searches"][format]["tree"]["queryGroup"]
+
+        if selected is True:
+            search["selectedtemplate"] = format
+
+        for spot in path:
+            [pos, gtype, static] = pathStepToPosGroupType(spot)
+            while len(curqry) <= pos:
+                curqry.append({})
+            if gtype is not None:
+                # This is a group
+                # If the inner node was not already set
+                if not curqry[pos]:
+                    curqry[pos]["pos"] = ""
+                    curqry[pos]["type"] = "group"
+                    curqry[pos]["val"] = gtype
+                    curqry[pos]["static"] = static
+                    curqry[pos]["queryGroup"] = []
+                # Move on to the next node in the path
+                curqry = curqry[pos]["queryGroup"]
+            else:
+                # This is a query
+
+                # Keep track of keys encountered
+                keys_seen = {}
+                for key in form_classes[format].form.base_fields.keys():
+                    keys_seen[key] = 0
+                cmpnts = []
+
+                curqry[pos]["type"] = "query"
+
+                # Set the form values in the query based on the form elements
+                for key in form.keys():
+                    # Remove "form-#-" from the form element ID
+                    cmpnts = key.split("-")
+                    keyname = cmpnts[-1]
+                    keys_seen[key] = 1
+                    if keyname == "pos":
+                        curqry[pos][key] = ""
+                    elif keyname == "static":
+                        if form[key] == "true":
+                            curqry[pos][key] = True
+                        else:
+                            curqry[pos][key] = False
+                    elif key not in curqry[pos]:
+                        curqry[pos][key] = form[key]
+                    else:
+                        # Log a warning
+                        print(
+                            f"WARNING: Unrecognized form element not set at pos {pos}: {key} to {form[key]}"
+                        )
+
+                # Now initialize anything missing a value to an empty string
+                # This is used to correctly reconstruct the user's query upon form_invalid
+                for key in form_classes[format].form.base_fields.keys():
+                    if keys_seen[key] == 0:
+                        curqry[pos][key] = ""
+    return search
+
+
+def pathStepToPosGroupType(spot):
+    """
+    Takes a substring from a pos field defining a single tree node and returns its position and group type (if it's an
+    inner node).  E.g. "0-all"
+    """
+
+    pos_gtype_stc = spot.split("-")
+    if len(pos_gtype_stc) == 3:
+        pos = pos_gtype_stc[0]
+        gtype = pos_gtype_stc[1]
+        if pos_gtype_stc[2] == "true":
+            static = True
+        else:
+            static = False
+    elif len(pos_gtype_stc) == 2:
+        pos = pos_gtype_stc[0]
+        gtype = pos_gtype_stc[1]
+        static = False
+    else:
+        pos = spot
+        gtype = None
+        static = False
+    pos = int(pos)
+    return [pos, gtype, static]
+
+
+def rootToFormatInfo(rootInfo):
+    """
+    Takes the first substring from a pos field defining the root node and returns the format code, format name, and
+    whether it is the selected format.
+    """
+
+    val_name_sel = rootInfo.split("-")
+    sel = False
+    name = ""
+    if len(val_name_sel) == 3:
+        val = val_name_sel[0]
+        name = val_name_sel[1]
+        if val_name_sel[2] == "selected":
+            sel = True
+    elif len(val_name_sel) == 2:
+        val = val_name_sel[0]
+        name = val_name_sel[1]
+    else:
+        print("WARNING: Unable to parse format name from submitted form data.")
+        val = val_name_sel
+        name = val_name_sel
+    return [val, name, sel]
